requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.1=pypi_0
altair=5.0.1=pypi_0
annotated-types=0.5.0=pypi_0
anyio=3.7.1=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.6=pypi_0
cmake=3.26.4=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.2=pypi_0
executing=1.2.0=pyhd8ed1ab_0
fastapi=0.100.1=pypi_0
ffmpy=0.3.1=pypi_0
filelock=3.12.2=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.41.0=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gradio=3.39.0=pypi_0
gradio-client=0.3.0=pypi_0
grpcio=1.56.0=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.16.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.0=pypi_0
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.12.0=pyh41d4057_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.1=pypi_0
jsonschema=4.18.4=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.12.0=py38h578d9bd_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
linkify-it-py=2.0.2=pypi_0
lit=16.0.6=pypi_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
onnx=1.14.0=pypi_0
onnx-simplifier=0.4.33=pypi_0
onnxconverter-common=1.13.0=pypi_0
onnxmltools=1.11.2=pypi_0
onnxruntime=1.15.1=pypi_0
opencv-python-headless=4.8.0.74=pypi_0
openssl=3.0.9=h7f8727e_0
orjson=3.9.2=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.0=pypi_0
pip=23.1.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.23.4=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydantic=2.1.1=pypi_0
pydantic-core=2.4.0=pypi_0
pydub=0.25.1=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-multipart=0.0.6=pypi_0
python_abi=3.8=2_cp38
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py38ha71036d_2
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.0=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.4.2=pypi_0
rpds-py=0.9.2=pypi_0
rsa=4.9=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=67.8.0=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
skl2onnx=1.14.1=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.27.0=pypi_0
sympy=1.12=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=pypi_0
torch=2.0.1=pypi_0
torchinfo=1.8.0=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.1=py38h0a891b7_3
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
triton=2.0.0=pypi_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023.3=pypi_0
uc-micro-py=1.0.2=pypi_0
urllib3=1.26.16=pypi_0
uvicorn=0.23.1=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
websockets=11.0.3=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
